signature errors to separate module
