+++ src/blockchain/transactions.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use exonum::blockchain::{ExecutionError, ExecutionResult, Transaction};
-use exonum::crypto::{Hash, PublicKey};
+use exonum::crypto::PublicKey;
 use exonum::helpers::ValidatorId;
 use exonum::messages::Message;
 use exonum::storage::Fork;
@@ -26,6 +26,7 @@ use btc;
 use BTC_ANCHORING_SERVICE_ID;
 
 use super::data_layout::TxInputId;
+use super::errors::SignatureError;
 use super::BtcAnchoringSchema;
 
 transactions! {
@@ -45,63 +46,6 @@ transactions! {
             /// Signature content.
             content: &[u8]
         }
-    }
-}
-
-#[derive(Debug, Fail)]
-pub enum SignatureError {
-    #[fail(
-        display = "Received signature for the incorrect anchoring transaction. Expected: {:?}. Received: {:?}.",
-        expected_id,
-        received_id
-    )]
-    Unexpected {
-        expected_id: Hash,
-        received_id: Hash,
-    },
-    #[fail(display = "Received signature for anchoring transaction while in transition state.")]
-    InTransition,
-    #[fail(display = "Public key of validator {:?} is missing.", _0)]
-    MissingPublicKey { validator_id: ValidatorId },
-    #[fail(display = "Input with index {} doesn't exist.", _0)]
-    NoSuchInput { idx: usize },
-    #[fail(display = "Signature verification failed.")]
-    VerificationFailed,
-    #[fail(display = "{}", _0)]
-    TxBuilderError(btc::BuilderError),
-    #[fail(display = "Unknown error")]
-    UnknownError,
-}
-
-#[derive(Debug)]
-pub enum ErrorCode {
-    Unexpected = 1,
-    InTransition = 2,
-    MissingPublicKey = 3,
-    NoSuchInput = 4,
-    VerificationFailed = 5,
-    TxBuilderError = 6,
-    UnknownError = 255,
-}
-
-impl SignatureError {
-    fn code(&self) -> ErrorCode {
-        match self {
-            SignatureError::Unexpected { .. } => ErrorCode::Unexpected,
-            SignatureError::InTransition => ErrorCode::InTransition,
-            SignatureError::MissingPublicKey { .. } => ErrorCode::MissingPublicKey,
-            SignatureError::NoSuchInput { .. } => ErrorCode::NoSuchInput,
-            SignatureError::VerificationFailed => ErrorCode::VerificationFailed,
-            SignatureError::TxBuilderError(..) => ErrorCode::TxBuilderError,
-            _ => ErrorCode::UnknownError,
-        }
-    }
-}
-
-impl From<SignatureError> for ExecutionError {
-    fn from(value: SignatureError) -> ExecutionError {
-        let description = format!("{}", value);
-        ExecutionError::with_description(value.code() as u8, description)
     }
 }
