+++ tests/mod.rs
@@ -1,3 +1,17 @@
+// Copyright 2018 The Exonum Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 extern crate bitcoin;
 extern crate exonum;
 extern crate exonum_btc_anchoring;
@@ -13,12 +27,11 @@ extern crate btc_transaction_utils;
 #[cfg(feature = "rpc_tests")]
 mod rpc_tests {
     use exonum::blockchain::{Transaction, TransactionErrorType};
-    use exonum::crypto::Hash;
     use exonum::explorer::BlockWithTransactions;
     use exonum::helpers::Height;
     use exonum_btc_anchoring::{
-        blockchain::transactions::ErrorCode, btc::BuilderError, config::GlobalConfig,
-        rpc::BtcRelay, test_data::AnchoringTestKit, BTC_ANCHORING_SERVICE_NAME,
+        blockchain::errors::ErrorCode, btc::BuilderError, config::GlobalConfig, rpc::BtcRelay,
+        test_data::AnchoringTestKit, BTC_ANCHORING_SERVICE_NAME,
     };
 
     fn assert_tx_error(block: BlockWithTransactions<Box<Transaction>>, e: ErrorCode) {
